build: Created makefile
Created basic makefile to find all source files in src/ folder, and
compile them. All source files are first analyzed to find out which
source files depend on which files. Then a real compilation step is
performed which compiles the source files and in the end links them
together.

The .gitignore is updated to also include the obj folder.